+++ MovieQuiz/Services/MovieQuizPresenter.swift
@@ -1,0 +1,32 @@
+//
+//  QuizConverter.swift
+//  MovieQuiz
+//
+//  Created by Konstantin Lyashenko on 18.03.2024.
+//
+
+import UIKit
+
+final class MovieQuizPresenter: QuizConverterProtocol {
+    let questionsAmount = 10
+    private var currentQuestionIndex = 0
+    
+    func convert(model: QuizQuestion) -> QuizStepViewModel {
+        let image = UIImage(data: model.image)
+        return QuizStepViewModel(image: image ?? UIImage(),
+                                 question: model.text,
+                                 questionNumber: "\(currentQuestionIndex)/\(questionsAmount)")
+    }
+    
+    func isLastQuestion() -> Bool {
+        currentQuestionIndex == questionsAmount - 1
+    }
+    
+    func resetQuestionIndex() {
+        currentQuestionIndex = 0
+    }
+    
+    func switchToNextQuestion() {
+        currentQuestionIndex += 1
+    }
+}
