--- MovieQuiz/Services/QuizConverter.swift
@@ -1,17 +1,0 @@
-//
-//  QuizConverter.swift
-//  MovieQuiz
-//
-//  Created by Konstantin Lyashenko on 18.03.2024.
-//
-
-import UIKit
-
-struct QuizConverter: QuizConverterProtocol {
-    static func convert(model: QuizQuestion, currentIndex: Int, totalCount: Int) -> QuizStepViewModel {
-        let image = UIImage(data: model.image)
-        return QuizStepViewModel(image: image ?? UIImage(),
-                                 question: model.text,
-                                 questionNumber: "\(currentIndex)/\(totalCount)")
-    }
-}
